pyproject.toml
[project]
name = "flask-vercel-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.3",
    "python-dotenv>=1.2.1",
    "supabase>=2.28.0",
]
